Fixed a bug in saving/loading and tweaked starting customer money.

+++ LemonadeStand/Player.cs
@@ -75,7 +75,7 @@ namespace LemonadeStand
         {
             moneySpentToday = 0;
         }
-        public void SetTotalSpent(int amount)
+        public void SetTotalSpent(double amount)
         {
             totalMoneySpent = amount;
         }
